requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
python-dotenv==1.0.1
unidiff==0.7.5
ollama==0.3.0
pydantic==2.7.4
